+++ lib/k8s/client/runner/base.ex
@@ -7,7 +7,7 @@ defmodule K8s.Client.Runner.Base do
           {:ok, map() | reference()}
           | {:error, K8s.Middleware.Error.t()}
           | {:error, :connection_not_registered}
-          | {:error, :missing_required_param}
+          | {:error, :missing_required_param, list(atom)}
           | {:error, atom()}
           | {:error, binary()}
 
@@ -85,72 +85,44 @@ defmodule K8s.Client.Runner.Base do
     do: run(conn, operation, [])
 
   @doc """
-  Run an operation and pass `opts` to HTTPoison.
+  Run an operation and pass `http_opts` to `K8s.Client.HTTPProvider`
   Destructures `Operation` data and passes as the HTTP body.
 
   See `run/2`
   """
   @spec run(Conn.t(), Operation.t(), keyword()) :: result_t
-  def run(%Conn{} = conn, %Operation{} = operation, opts) when is_list(opts) do
-    run(conn, operation, operation.data, opts)
+  def run(%Conn{} = conn, %Operation{} = operation, http_opts) when is_list(http_opts) do
+    run(conn, operation, operation.data, http_opts)
   end
 
   @doc """
-  Run an operation with an HTTP Body (map) and pass `opts` to HTTPoison.
+  Run an operation with an HTTP Body (map) and pass `http_opts` to `K8s.Client.HTTPProvider`.
   See `run/2`
   """
   @spec run(Conn.t(), Operation.t(), map(), keyword()) :: result_t
-  def run(%Conn{} = conn, %Operation{} = operation, body, opts \\ []) do
+  def run(%Conn{} = conn, %Operation{} = operation, body, http_opts \\ []) do
     with {:ok, url} <- K8s.Discovery.url_for(conn, operation),
-         req <- new_request(conn, url, operation, body, opts),
+         req <- new_request(conn, url, operation, body, http_opts),
          {:ok, req} <- K8s.Middleware.run(req) do
       conn.http_provider.request(req.method, req.url, req.body, req.headers, req.opts)
     end
   end
 
-  @spec new_request(Conn.t(), String.t(), K8s.Operation.t(), body_t, Keyword.t()) ::
+  @spec new_request(Conn.t(), String.t(), Operation.t(), body_t, Keyword.t()) ::
           Request.t()
-  defp new_request(%Conn{} = conn, url, %Operation{} = operation, body, opts) do
-    req = %Request{conn: conn, method: operation.method, body: body}
+  defp new_request(%Conn{} = conn, url, %Operation{} = operation, body, http_opts) do
+    req = %Request{conn: conn, method: operation.method, body: body, url: url}
 
-    params = merge_deprecated_params(operation.query_params, opts[:params])
+    query_params = build_query_params(operation)
+    http_opts_with_selector_params = Keyword.put(http_opts, :params, query_params)
+    updated_http_opts = Keyword.merge(req.opts, http_opts_with_selector_params)
 
-    # if label_selector is set, the end user is setting it manually, respect that value until removed in #73
-    label_selector =
-      maybe_get_deprecated_label_selector(params[:labelSelector], operation.label_selector)
-
-    http_opts_params = build_http_params(params, label_selector)
-
-    opts_with_selector_params = Keyword.put(opts, :params, http_opts_params)
-    http_opts = Keyword.merge(req.opts, opts_with_selector_params)
-
-    %Request{req | opts: http_opts, url: url}
+    %Request{req | opts: updated_http_opts}
   end
 
-  @spec maybe_get_deprecated_label_selector(K8s.Selector.t() | nil, K8s.Selector.t() | nil) ::
-          K8s.Selector.t() | nil
-  defp maybe_get_deprecated_label_selector(new_label_selector, nil), do: new_label_selector
-
-  @deprecated "K8s.Operation label_selector is deprecated. Use K8s.Selector functions instead."
-  defp maybe_get_deprecated_label_selector(nil, deprecated_label_selector),
-    do: deprecated_label_selector
-
-  @spec merge_deprecated_params(map(), nil | map()) :: map()
-  defp merge_deprecated_params(op_params, nil), do: op_params
-
-  @deprecated "Providing HTTPoison options to K8s.Client.Runner.Base.run/N is deprecated. Use K8s.Operation's query_params key instead."
-  defp merge_deprecated_params(%{} = op_params, run_params) do
-    run_params_as_map = Enum.into(run_params, %{})
-    Map.merge(op_params, run_params_as_map)
-  end
-
-  @spec build_http_params(keyword | map, nil | K8s.Selector.t()) :: map()
-  defp build_http_params(params, nil), do: Enum.into(params, %{})
-
-  # Supplying a `labelSelector` to `run/4 should take precedence
-  defp build_http_params(params, %K8s.Selector{} = s) do
-    # After HTTPoison options are removed from run/N, this will always be a map()
-    params_as_map = Enum.into(params, %{})
-    Map.merge(params_as_map, %{labelSelector: K8s.Selector.to_s(s)})
+  @spec build_query_params(Operation.t()) :: keyword()
+  defp build_query_params(%Operation{} = operation) do
+    label_selector = Operation.get_label_selector(operation)
+    Keyword.merge(operation.query_params, labelSelector: K8s.Selector.to_s(label_selector))
   end
 end